ci(e2e): run AWS, DO, Google Cloud e2e tests on main branch push
This was triggered by our EKS deployment starting to fail due to a
[networking
change](https://github.com/PostHog/charts-clickhouse/pull/267), but we
did not pick up the issue for some time.

It would be ideal that we run these on PRs also, but I've gone for just
main for now such that we at least know when main it broken. I think
some optimizations might be in order before these are enabled on PRs,
contraints being we are using let's encrypt and I'm not sure if e.g.
hitting rate limits here is likely. Maybe this is too defensive?

Other options for PRs: run on github comment etc.