fix(ingress-nginx): forward X-Forwarded-* headers to upstreams
Within the posthog app we use the X-Forwarded-* headers to, for
instance, determine the scheme to use when building absolute urls.
Previously we were not setting X-Forwarded-* headers hence we ended up
producing http scheme urls, which causes issue, among other things, with
cross domain requests.

Here we set on the nginx ingress controller that we should always
forward these headers. We could also have added these as an annotation
to the ingress resources as a snippet, but I've just gone for applying
this across the board.

This was as a result of debugging
https://posthog.slack.com/archives/C02G72Y6E9H/p1646072061024549