fix(clickhouse-operator): pin docker image version (#371)
* fix(clickhouse-operator): pin docker image version

Previously we were using the latest available image. Here we pin to the
latest image.

It looks like they version the chart and the operator image the same. It
may be worth pinning them to the same version. However, this will cause
the operator to downgrade in existing installs, so here I'm opting to
try to keep the image version the same as current installs.

* update snapshot;

* Add comment

* fix