feat: add ability to specify existing secret for Django SECRET_KEY
Previously the behaviour was to look for an existing value for the
secret and, if one doesn't exist, to generate one automatically.

There wasn't a way to explicitly specify the key, which means that if
you want to point a new deployment at existing database resources,
you'll end up invalidating e.g. Django issued cookies.