chore: remove unique-name secret
It doesn't appear to be being used.

As a result of adding the ability to specify your own `SECRET_KEY` via
`posthogSecretKey.existingKey` it is now the case that if you also set
`email.existingKey` your'll end up with `data == null` within the
`posthog` secret. This causes issues with the evaluation of `index
$previous.data 'unique-name'`

Rather than resolving this, I've opted for removal due to it not being
used.