feat(hpa): add non-abstracted hpa config values
Previously we had a thin abstraction over a subset of the hpa config.
Rather than maintain this abstraction I have simply exposed the
underlying spec for hpa manifests. This was we can use the full range of
hpa without needing to make chart interface changes.

I've added this as the existing configuration can exhibit a lot of
flapping, where we might go from 10 pods, down to 5, only to then
request another 5 pods directly after. I would like to set the
`behavior.scaleDown.stabilizationWindowSeconds` setting to a higher
value than the default.

Note that this does not include the pluginsAsync values addition at the
moment. This is because we do not have an HPA manifest for pluginsAsync,
I'll handle this separately.

The reason for not reusing e.g. `web.hpa` is to avoid any unexpected
behaviour with existing values people may have. I chose `autoscaling`
bsaed on a similarly structured value in the `ingress-nginx` helm chart.