feat(logging): log Kubernetes with eventrouter
We currently can only view events via `kubectl get events` which isn't
great for correlating with logs. At the moment I'm trying to understand
why some requests hit 502 on deploys, which appears to be due to the
pods being killed without grace, but it's difficult to see exactly what
is going on at the moment.

I followed the instructions on
https://grafana.com/blog/2020/07/21/loki-tutorial-how-to-send-logs-from-eks-with-promtail-to-get-full-visibility-in-grafana/

There is also
https://grafana.com/docs/agent/latest/configuration/integrations/integrations-next/eventhandler-config/
which we should perhaps move to but for now, this should do the trick.