fix(web): wait 10 seconds for k8s endpoints to be updated on term
To ensure that we initiate Gunicorn shutdown after the last web requests
come in via the Service endpoint, we wait 10 seconds thereby giving time
to have the Pod removed from the Endpoint.

We are seeing 502s on deploys. Gunicorn is successfully waiting to
complete in flight requests now, but will reject any new ones that come
in after shutdown has started.