feat(ingress): allow setting ingress.hostname to empty (#568)
* feat(ingress): allow setting ingress.hostname to empty

Previously you would be able to set ingress.hostname to an empty string.
However, the SITE_URL would be set incorrectly, thus e.g. links to
sharable dashboards would no longer work.

We want to be able to set the hostname to empty such that the ingress
controller will pick up any traffic, no matter what the "Host" http
header. This was we can allow users to provide reverse proxies on
different addresses which are just proxying the original Host headers
through.

* don't raise on hostname not being set

I added some security in to ensure that the `SITE_URL` env would be set
accordingly. However, that results in existing tests failing and would
likely be backwards incompat. for lots of deployments.

* typo