ci(monitoring): install loki/promtail/grafana in CI k3s deployments
I was trying to [upgrade Loki Helm
chart](https://github.com/PostHog/charts-clickhouse/pull/570) to version
3.x (which is largely the previous loki-simple-scalable chart), which
would allow us to scale for more log bandwidth.

However, it requires manual installing of CRDs if we are upgrading.
Adding this to the k3s tests means we will also be testing the upgrade
and would have spotted this.

It does however mean that if there are any flakes caused by e.g.
additional memory requirements or another factor then it could be
annoying to have these e.g. block a release.