refactor(plugins-server): templatize plugin server deployment (#598)
* refactor(plugins-server): templatize plugin server deployment

This is a precursor to further splitting the plugin server. It is a
little heavy handed and maybe an over abstraction but we'll see. Should
be easy to back out with the unittesting in place.

* chore: include hpa in pluginserver abstraction

* add new line

* add new lines

* Separate out plugin server event more

* fix unittest

* wip

* wip

* switch to async-handlers

* wip

* revert to old deployments in tests, we cant easily test the new image

* ignore restarts in pgbouncer

* fix mode setting

* format

* chore: update async-handlers -> exports

* chore: update Chart version by a minor version

This release should be backwards compatible and only adds new
functionality for allowing to further separate the plugin workloads.