feat(alerts): add alert on exports consumer group backlog
Adds a critical alert that checks to see it the delta in consumer lag
between now and 10 minutes ago is consistently >0 for 10 minutes
straight.

There are possibly cases where this doesn't work. I did the detla as I
didn't want to add an arbitrary message backlog value that might not
translate to all deployments of the app.

This should however catch the case were the lag is stribtly
monotonically increasing.