fix(ingress): ensure only capture and decide go to events deployment
Previously we would match e.g. any path that started with `/e` and route
this to the `events` pods. This isn't desired as this captures e.g.
`/events/` which is a user facing UI. Likewise for the other endpoints
that were previously k8s ingress `Prefix` rules.

Note that the ingress rules only match on path i.e. before any query
string components and the `?` character.