chore(bitnami): use the pre-2022 bitnami repo index.yaml (#658)
Bitnami did a bit of a size reduction of their index.yaml for their Helm
repo in the last day or so: https://github.com/bitnami/charts/pull/14024

We previously changed to using pre-2022 for other charts that fell out
of the [6 month retention
policy](https://github.com/bitnami/charts/issues/10539) but not for all.
Looks like a few more have now dropped on of the main index so I'm
changing all to use the github hosted pre-2022 index.

We should actually update these charts at some point, but given these
will likely be major upgrades I'm not doing it for now. My main aim is
to get back to a point where we can release our chart, currently CI
chart-releaser is failing.