chore(session-recordings): separate session recordings from events
We currently have session recordings events going to the same endpoint
as analytics events and decide endpiont. The characteristics of session
recordings is quite different, and our expectations of priority. It's
much more important that we capture events and service decide endpoints
than we capture all session recordings.

This separates the deployment to offer more flexibility in providing
different service availability.