chore: Make Celery Queue Depth page less (#794)
companion to https://github.com/PostHog/charts/pull/773

We've been paging for this daily for more than 10 days. In all of those cases there hasn't been any action and it resolves itself. This creates alert fatigue, so let's make the alert less sensitive.
We can additionally add a non-paging alert, if this is deemed useful.